moved pages navbar to /pages
