+++ src/pages/GalleryPage/index.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import PagesNavbar from "../../components/PagesNavbar";
+import PagesNavbar from "../PagesNavbar";
 import { PhotoAlbum } from "react-photo-album";
 import Lightbox from "yet-another-react-lightbox";
 import "yet-another-react-lightbox/styles.css";
